Refactor: CardContext 수정
