+++ src/main.tsx
@@ -6,7 +6,7 @@ import { createRoot } from 'react-dom/client';
 import App from './App';
 import { RouteProvider } from './contexts';
 import Portal from './components/Portal';
-import { CardProvider } from './templates/CardAddPage';
+import { CardProvider } from './contexts/CardContext';
 
 createRoot(document.getElementById('root') as HTMLElement).render(
   <StrictMode>
