+++ src/templates/CardAddPage/index.ts
@@ -1,1 +1,1 @@
-export * from './CardAddPage.context';
+export {};
